+++ simapp/app.go
@@ -384,8 +384,8 @@ func NewSimApp(
 		upgrade.NewAppModule(app.UpgradeKeeper),
 		evidence.NewAppModule(app.EvidenceKeeper),
 		params.NewAppModule(app.ParamsKeeper),
-		tokenmodule.NewAppModule(appCodec, app.TokenKeeper),
-		collectionmodule.NewAppModule(appCodec, app.CollectionKeeper),
+		tokenmodule.NewAppModule(appCodec, app.TokenKeeper, app.AccountKeeper),
+		collectionmodule.NewAppModule(appCodec, app.CollectionKeeper, app.AccountKeeper),
 		authzmodule.NewAppModule(appCodec, app.AuthzKeeper, app.AccountKeeper, app.BankKeeper, app.interfaceRegistry),
 	)