+++ x/collection/keeper/grpc_query.go
@@ -17,15 +17,59 @@ import (
 )
 
 type queryServer struct {
-	keeper Keeper
+	keeper     Keeper
+	authKeeper collection.AuthKeeper
 }
 
 // NewQueryServer returns an implementation of the token QueryServer interface
 // for the provided Keeper.
-func NewQueryServer(keeper Keeper) collection.QueryServer {
+func NewQueryServer(keeper Keeper, authKeeper collection.AuthKeeper) collection.QueryServer {
 	return &queryServer{
-		keeper: keeper,
+		keeper:     keeper,
+		authKeeper: authKeeper,
 	}
+}
+
+func (s queryServer) validateExistenceOfAccountGRPC(ctx sdk.Context, addr sdk.AccAddress) error {
+	if !s.authKeeper.HasAccount(ctx, addr) {
+		return status.Error(codes.NotFound, sdkerrors.ErrUnknownAddress.Wrap(addr.String()).Error())
+	}
+
+	return nil
+}
+
+func (s queryServer) validateExistenceOfCollectionGRPC(ctx sdk.Context, id string) error {
+	if _, err := s.keeper.GetContract(ctx, id); err != nil {
+		return status.Error(codes.NotFound, err.Error())
+	}
+
+	return nil
+}
+
+func (s queryServer) validateExistenceOfFTClassGRPC(ctx sdk.Context, contractID, classID string) error {
+	class, err := s.keeper.GetTokenClass(ctx, contractID, classID)
+	if err != nil {
+		return status.Error(codes.NotFound, err.Error())
+	}
+
+	_, ok := class.(*collection.FTClass)
+	if !ok {
+		return status.Error(codes.NotFound, sdkerrors.ErrInvalidType.Wrapf("not a class of fungible token: %s", classID).Error())
+	}
+	return nil
+}
+
+func (s queryServer) validateExistenceOfNFTClassGRPC(ctx sdk.Context, contractID, classID string) error {
+	class, err := s.keeper.GetTokenClass(ctx, contractID, classID)
+	if err != nil {
+		return status.Error(codes.NotFound, err.Error())
+	}
+
+	_, ok := class.(*collection.NFTClass)
+	if !ok {
+		return status.Error(codes.NotFound, sdkerrors.ErrInvalidType.Wrapf("not a class of non-fungible token: %s", classID).Error())
+	}
+	return nil
 }
 
 var _ collection.QueryServer = queryServer{}
@@ -50,6 +94,15 @@ func (s queryServer) Balance(c context.Context, req *collection.QueryBalanceRequ
 	}
 
 	ctx := sdk.UnwrapSDKContext(c)
+
+	if err := s.validateExistenceOfAccountGRPC(ctx, addr); err != nil {
+		return nil, err
+	}
+
+	if err := s.validateExistenceOfCollectionGRPC(ctx, req.ContractId); err != nil {
+		return nil, err
+	}
+
 	balance := s.keeper.GetBalance(ctx, req.ContractId, addr, req.TokenId)
 	coin := collection.Coin{
 		TokenId: req.TokenId,
@@ -113,9 +166,15 @@ func (s queryServer) FTSupply(c context.Context, req *collection.QueryFTSupplyRe
 	classID := collection.SplitTokenID(req.TokenId)
 
 	ctx := sdk.UnwrapSDKContext(c)
-	if _, err := s.keeper.GetTokenClass(ctx, req.ContractId, classID); err != nil {
+
+	if err := s.validateExistenceOfCollectionGRPC(ctx, req.ContractId); err != nil {
 		return nil, err
 	}
+
+	if err := s.validateExistenceOfFTClassGRPC(ctx, req.ContractId, classID); err != nil {
+		return nil, err
+	}
+
 	supply := s.keeper.GetSupply(ctx, req.ContractId, classID)
 
 	return &collection.QueryFTSupplyResponse{Supply: supply}, nil
@@ -137,9 +196,15 @@ func (s queryServer) FTMinted(c context.Context, req *collection.QueryFTMintedRe
 	classID := collection.SplitTokenID(req.TokenId)
 
 	ctx := sdk.UnwrapSDKContext(c)
-	if _, err := s.keeper.GetTokenClass(ctx, req.ContractId, classID); err != nil {
+
+	if err := s.validateExistenceOfCollectionGRPC(ctx, req.ContractId); err != nil {
 		return nil, err
 	}
+
+	if err := s.validateExistenceOfFTClassGRPC(ctx, req.ContractId, classID); err != nil {
+		return nil, err
+	}
+
 	minted := s.keeper.GetMinted(ctx, req.ContractId, classID)
 
 	return &collection.QueryFTMintedResponse{Minted: minted}, nil
@@ -161,9 +226,15 @@ func (s queryServer) FTBurnt(c context.Context, req *collection.QueryFTBurntRequ
 	classID := collection.SplitTokenID(req.TokenId)
 
 	ctx := sdk.UnwrapSDKContext(c)
-	if _, err := s.keeper.GetTokenClass(ctx, req.ContractId, classID); err != nil {
+
+	if err := s.validateExistenceOfCollectionGRPC(ctx, req.ContractId); err != nil {
 		return nil, err
 	}
+
+	if err := s.validateExistenceOfFTClassGRPC(ctx, req.ContractId, classID); err != nil {
+		return nil, err
+	}
+
 	burnt := s.keeper.GetBurnt(ctx, req.ContractId, classID)
 
 	return &collection.QueryFTBurntResponse{Burnt: burnt}, nil
@@ -184,9 +255,15 @@ func (s queryServer) NFTSupply(c context.Context, req *collection.QueryNFTSupply
 	}
 
 	ctx := sdk.UnwrapSDKContext(c)
-	if _, err := s.keeper.GetTokenClass(ctx, req.ContractId, classID); err != nil {
+
+	if err := s.validateExistenceOfCollectionGRPC(ctx, req.ContractId); err != nil {
 		return nil, err
 	}
+
+	if err := s.validateExistenceOfNFTClassGRPC(ctx, req.ContractId, classID); err != nil {
+		return nil, err
+	}
+
 	supply := s.keeper.GetSupply(ctx, req.ContractId, classID)
 
 	return &collection.QueryNFTSupplyResponse{Supply: supply}, nil
@@ -207,9 +284,15 @@ func (s queryServer) NFTMinted(c context.Context, req *collection.QueryNFTMinted
 	}
 
 	ctx := sdk.UnwrapSDKContext(c)
-	if _, err := s.keeper.GetTokenClass(ctx, req.ContractId, classID); err != nil {
+
+	if err := s.validateExistenceOfCollectionGRPC(ctx, req.ContractId); err != nil {
 		return nil, err
 	}
+
+	if err := s.validateExistenceOfNFTClassGRPC(ctx, req.ContractId, classID); err != nil {
+		return nil, err
+	}
+
 	minted := s.keeper.GetMinted(ctx, req.ContractId, classID)
 
 	return &collection.QueryNFTMintedResponse{Minted: minted}, nil
@@ -230,9 +313,15 @@ func (s queryServer) NFTBurnt(c context.Context, req *collection.QueryNFTBurntRe
 	}
 
 	ctx := sdk.UnwrapSDKContext(c)
-	if _, err := s.keeper.GetTokenClass(ctx, req.ContractId, classID); err != nil {
+
+	if err := s.validateExistenceOfCollectionGRPC(ctx, req.ContractId); err != nil {
 		return nil, err
 	}
+
+	if err := s.validateExistenceOfNFTClassGRPC(ctx, req.ContractId, classID); err != nil {
+		return nil, err
+	}
+
 	burnt := s.keeper.GetBurnt(ctx, req.ContractId, classID)
 
 	return &collection.QueryNFTBurntResponse{Burnt: burnt}, nil
@@ -250,7 +339,7 @@ func (s queryServer) Contract(c context.Context, req *collection.QueryContractRe
 	ctx := sdk.UnwrapSDKContext(c)
 	contract, err := s.keeper.GetContract(ctx, req.ContractId)
 	if err != nil {
-		return nil, err
+		return nil, status.Error(codes.NotFound, err.Error())
 	}
 
 	return &collection.QueryContractResponse{Contract: *contract}, nil
@@ -271,9 +360,14 @@ func (s queryServer) TokenClassTypeName(c context.Context, req *collection.Query
 	}
 
 	ctx := sdk.UnwrapSDKContext(c)
+
+	if err := s.validateExistenceOfCollectionGRPC(ctx, req.ContractId); err != nil {
+		return nil, err
+	}
+
 	class, err := s.keeper.GetTokenClass(ctx, req.ContractId, req.ClassId)
 	if err != nil {
-		return nil, err
+		return nil, status.Error(codes.NotFound, err.Error())
 	}
 	name := proto.MessageName(class)
 
@@ -297,12 +391,12 @@ func (s queryServer) TokenType(c context.Context, req *collection.QueryTokenType
 	ctx := sdk.UnwrapSDKContext(c)
 	class, err := s.keeper.GetTokenClass(ctx, req.ContractId, classID)
 	if err != nil {
-		return nil, err
+		return nil, status.Error(codes.NotFound, err.Error())
 	}
 
 	nftClass, ok := class.(*collection.NFTClass)
 	if !ok {
-		return nil, sdkerrors.ErrInvalidType.Wrapf("not a class of non-fungible token: %s", classID)
+		return nil, status.Error(codes.NotFound, sdkerrors.ErrInvalidType.Wrapf("not a class of non-fungible token: %s", classID).Error())
 	}
 
 	tokenType := collection.TokenType{
@@ -372,7 +466,7 @@ func (s queryServer) Token(c context.Context, req *collection.QueryTokenRequest)
 	ctx := sdk.UnwrapSDKContext(c)
 	legacyToken, err := s.getToken(ctx, req.ContractId, req.TokenId)
 	if err != nil {
-		return nil, err
+		return nil, status.Error(codes.NotFound, err.Error())
 	}
 
 	any, err := codectypes.NewAnyWithValue(legacyToken)
@@ -397,8 +491,13 @@ func (s queryServer) Root(c context.Context, req *collection.QueryRootRequest) (
 	}
 
 	ctx := sdk.UnwrapSDKContext(c)
-	if err := s.keeper.hasNFT(ctx, req.ContractId, req.TokenId); err != nil {
+
+	if err := s.validateExistenceOfCollectionGRPC(ctx, req.ContractId); err != nil {
 		return nil, err
+	}
+
+	if err := s.keeper.hasNFT(ctx, req.ContractId, req.TokenId); err != nil {
+		return nil, status.Error(codes.NotFound, err.Error())
 	}
 
 	root := s.keeper.GetRoot(ctx, req.ContractId, req.TokenId)
@@ -424,8 +523,13 @@ func (s queryServer) Parent(c context.Context, req *collection.QueryParentReques
 	}
 
 	ctx := sdk.UnwrapSDKContext(c)
-	if err := s.keeper.hasNFT(ctx, req.ContractId, req.TokenId); err != nil {
+
+	if err := s.validateExistenceOfCollectionGRPC(ctx, req.ContractId); err != nil {
 		return nil, err
+	}
+
+	if err := s.keeper.hasNFT(ctx, req.ContractId, req.TokenId); err != nil {
+		return nil, status.Error(codes.NotFound, err.Error())
 	}
 
 	parent, err := s.keeper.GetParent(ctx, req.ContractId, req.TokenId)
@@ -455,8 +559,13 @@ func (s queryServer) Children(c context.Context, req *collection.QueryChildrenRe
 	}
 
 	ctx := sdk.UnwrapSDKContext(c)
-	if err := s.keeper.hasNFT(ctx, req.ContractId, req.TokenId); err != nil {
+
+	if err := s.validateExistenceOfCollectionGRPC(ctx, req.ContractId); err != nil {
 		return nil, err
+	}
+
+	if err := s.keeper.hasNFT(ctx, req.ContractId, req.TokenId); err != nil {
+		return nil, status.Error(codes.NotFound, err.Error())
 	}
 
 	store := ctx.KVStore(s.keeper.storeKey)
@@ -494,6 +603,15 @@ func (s queryServer) GranteeGrants(c context.Context, req *collection.QueryGrant
 	}
 
 	ctx := sdk.UnwrapSDKContext(c)
+
+	if err := s.validateExistenceOfAccountGRPC(ctx, granteeAddr); err != nil {
+		return nil, err
+	}
+
+	if err := s.validateExistenceOfCollectionGRPC(ctx, req.ContractId); err != nil {
+		return nil, err
+	}
+
 	store := ctx.KVStore(s.keeper.storeKey)
 	grantStore := prefix.NewStore(store, grantKeyPrefixByGrantee(req.ContractId, granteeAddr))
 	var grants []collection.Grant
@@ -531,6 +649,18 @@ func (s queryServer) IsOperatorFor(c context.Context, req *collection.QueryIsOpe
 	}
 
 	ctx := sdk.UnwrapSDKContext(c)
+
+	if err := s.validateExistenceOfAccountGRPC(ctx, operator); err != nil {
+		return nil, err
+	}
+	if err := s.validateExistenceOfAccountGRPC(ctx, holder); err != nil {
+		return nil, err
+	}
+
+	if err := s.validateExistenceOfCollectionGRPC(ctx, req.ContractId); err != nil {
+		return nil, err
+	}
+
 	_, err = s.keeper.GetAuthorization(ctx, req.ContractId, holder, operator)
 	authorized := (err == nil)