+++ x/collection/keeper/keeper_test.go
@@ -64,7 +64,7 @@ func (s *KeeperTestSuite) SetupTest() {
 	s.goCtx = sdk.WrapSDKContext(s.ctx)
 	s.keeper = app.CollectionKeeper
 
-	s.queryServer = keeper.NewQueryServer(s.keeper)
+	s.queryServer = keeper.NewQueryServer(s.keeper, app.AccountKeeper)
 	s.msgServer = keeper.NewMsgServer(s.keeper)
 
 	s.depthLimit = 4
@@ -81,6 +81,10 @@ func (s *KeeperTestSuite) SetupTest() {
 	}
 	for i, address := range createRandomAccounts(len(addresses)) {
 		*addresses[i] = address
+
+		// create account
+		acc := app.AccountKeeper.NewAccountWithAddress(s.ctx, address)
+		app.AccountKeeper.SetAccount(s.ctx, acc)
 	}
 
 	s.balance = sdk.NewInt(1000000)