+++ CHANGELOG.md
@@ -59,6 +59,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Breaking Changes
 * (proto) [\#923](https://github.com/line/lbm-sdk/pull/923) deprecate broadcast mode `block`
+* (x/collection,token) [\#956](https://github.com/line/lbm-sdk/pull/956) Replace query errors on the original modules into gRPC ones
 
 ### Build, CI
 * (ci, build) [\#901](https://github.com/line/lbm-sdk/pull/901) Update release pipeline to match non-wasm env